moding experiment module to a file of itsown
